+++ CHANGELOG.md
@@ -20,6 +20,9 @@
 * Setting: Updates to Action profiles (UIDATIMP-1026)
 * Adjust the FOLIO record type in the match profile (UIDATIMP-1029)
 * Settings : Updates to default SRS Holdings job profile (UIDATIMP-1023)
+* Populate Record Log title column when imported record is MARC Holdings (UIDATIMP-1007)
+* Cover `<Jobs>` component with tests (UIDATIMP-959)
+* Cover `<UploadingJobsContextProvider>` component with tests (UIDATIMP-729)
 * Cover `<MatchProfiles>` settings with tests (UIDATIMP-975)
 
 ## [5.0.1](https://github.com/folio-org/ui-data-import/tree/v5.0.1) (2021-10-19)